INFRA-2103 Update build-and-push docker image workflow
While building a new image for the Cognito custom sender Lambda function
using this workflow it's been found that Lambda does not yet support
multi-arch container images.
That is, in order to reuse this GH workflow for the Lambda function
image creation, an ability to change platforms to match the compute
architecture and disable provenance should be added to the workflow.
This change adds additional inputs to the workflow and sets the default
values to match the current behavior.